backend/requirements.txt
# Web framework
Flask==3.1.0
Flask-RESTful==0.3.10

# Database
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
psycopg2-binary==2.9.10

# Authentication
Flask-JWT-Extended==4.7.1
Werkzeug==3.1.3

# API documentation
Flasgger==0.9.7.1

# Environment variables
python-dotenv==1.1.0

# Testing
pytest==8.3.5
pytest-flask==1.3.0

# CORS (allows React frontend to call the API)
Flask-CORS==5.0.1

# Anthropic AI — chatbot backend
anthropic==0.49.0

# RAG 
chromadb
sentence-transformers